separate FeatureFlagDelegate and FeatureFlagManager files
