+++ src/main/java/com/mixpanel/android/mpmetrics/FeatureFlagDelegate.java
@@ -1,0 +1,14 @@
+package com.mixpanel.android.mpmetrics;
+
+import org.json.JSONObject;
+
+/**
+ * Interface for FeatureFlagManager to retrieve necessary data and trigger actions
+ * from the main MixpanelAPI instance.
+ */
+interface FeatureFlagDelegate {
+    MPConfig getMPConfig();
+    String getDistinctId();
+    void track(String eventName, JSONObject properties);
+    String getToken();
+}
